+++ contrib/opencensus-ext-httplib/CHANGELOG.md
@@ -5,7 +5,7 @@
 ## 0.7.3
 Released 2020-02-03
 
-- Support span attribute changes in `opencensus>=0.7.7`
+- Added `component` span attribute
 
 ## 0.7.2
 Released 2019-08-06